requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
watchfiles>=0.19.0
flask>=3.0.0
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
bcrypt==4.1.1
cryptography>=41.0.0
requests>=2.28.0
sqlalchemy>=2.0.0
tushare==1.2.89
akshare>=1.17.65
baostock>=0.8.9
jqdatasdk>=1.8.16
python-multipart==0.0.6
yfinance==0.2.66
alipay-sdk-python>=3.3.398
wechatpayv3>=1.2.16
qrcode[pil]>=7.4
